split TestCenter controller into proctor specific and manager
