+++ controller/Delivery.php
@@ -33,10 +33,8 @@ use oat\taoProctoring\helpers\ReasonCategory;
  * @license GPL-2.0
  *
  */
-class Delivery extends \tao_actions_CommonModule implements ProctoringInterface
+class Delivery extends ProctoringModule
 {
-    use ProctoringTrait;
-
     /**
      * Displays the index page of the deliveries list all available deliveries for the current test center
      */
