+++ controller/TestCenter.php
@@ -23,7 +23,6 @@ namespace oat\taoProctoring\controller;
 
 use oat\taoProctoring\helpers\Breadcrumbs;
 use oat\taoProctoring\helpers\TestCenter as TestCenterHelper;
-use oat\taoProctoring\model\TestCenterService;
 
 /**
  * Proctoring Test Center controllers for test center screens
@@ -33,67 +32,8 @@ use oat\taoProctoring\model\TestCenterService;
  * @license GPL-2.0
  *
  */
-class TestCenter extends \tao_actions_SaSModule implements ProctoringInterface
+class TestCenter extends ProctoringModule
 {
-    use ProctoringTrait;
-
-    /**
-     * Initialize the service and the default data
-     */
-    public function __construct()
-    {
-        parent::__construct();
-        $this->service = $this->getClassService();
-    }
-
-    protected function getClassService()
-    {
-        return TestCenterService::singleton();
-    }
-
-    /**
-     * Edit a Test Center instance
-     * @return void
-     */
-    public function editCenter()
-    {
-        $clazz = $this->getCurrentClass();
-        $testCenter = $this->getCurrentInstance();
-
-        $formContainer = new \tao_actions_form_Instance($clazz, $testCenter);
-        $myForm = $formContainer->getForm();
-        if ($myForm->isSubmited()) {
-            if ($myForm->isValid()) {
-
-                $binder = new \tao_models_classes_dataBinding_GenerisFormDataBinder($testCenter);
-                $testCenter = $binder->bind($myForm->getValues());
-
-                $this->setData("selectNode", \tao_helpers_Uri::encode($testCenter->getUri()));
-                $this->setData('message', __('Test center saved'));
-                $this->setData('reload', true);
-            }
-        }
-
-        $memberProperty = new \core_kernel_classes_Property(TestCenterService::PROPERTY_MEMBERS_URI);
-        $memberForm = \tao_helpers_form_GenerisTreeForm::buildReverseTree($testCenter, $memberProperty);
-        $memberForm->setData('title', __('Select test-takers for the test center'));
-        $this->setData('memberForm', $memberForm->render());
-
-        $groupProperty = new \core_kernel_classes_Property(TestCenterService::PROPERTY_DELIVERY_URI);
-        $groupForm = \tao_helpers_form_GenerisTreeForm::buildTree($testCenter, $groupProperty);
-        $groupForm->setData('title', __('Select deliveries available at the test center'));
-        $this->setData('groupForm', $groupForm->render());
-
-        $proctorProperty = new \core_kernel_classes_Property(TestCenterService::PROPERTY_PROCTORS_URI);
-        $proctorForm = \tao_helpers_form_GenerisTreeForm::buildReverseTree($testCenter, $proctorProperty);
-        $proctorForm->setData('title', __('Select proctors for the test center'));
-        $this->setData('proctorForm', $proctorForm->render());
-
-        $this->setData('formTitle', __('Edit test center'));
-        $this->setData('myForm', $myForm->render());
-        $this->setView('form_test_center.tpl');
-    }
-
     /**
      * Displays the index page of the extension: list all available deliveries.
      */
