+++ controller/structures.xml
@@ -5,12 +5,12 @@
         <description>Test center contains test taker and groups and are administer by a proctor</description>
         <icon id="icon-test-takers" src=""/>
         <sections>
-            <section id="manage_testcenter" name="Manage Test Center" url="/taoProctoring/TestCenter/index">
+            <section id="manage_testcenter" name="Manage Test Center" url="/taoProctoring/TestCenterManager/index">
                 <trees>
                     <tree name="Test Center List"
                           className="Test Center"
                           rootNode="http://www.tao.lu/Ontologies/TAOTestCenter.rdf#TestCenter"
-                          dataUrl="/taoProctoring/TestCenter/getOntologyData"
+                          dataUrl="/taoProctoring/TestCenterManager/getOntologyData"
                           selectClass="testcenter-class-properties"
                           selectInstance="testcenter-properties"
                           moveInstance="testcenter-move"
@@ -22,26 +22,26 @@
                             group="content" context="class">
                         <icon id="icon-edit"/>
                     </action>
-                    <action id="testcenter-properties" name="Properties" url="/taoProctoring/TestCenter/editCenter"
+                    <action id="testcenter-properties" name="Properties" url="/taoProctoring/TestCenterManager/editCenter"
                             group="content" context="instance">
                         <icon id="icon-edit"/>
                     </action>
                     <action id="testcenter-filter" name="Filter" context="*" group="search" binding="filter">
                         <icon id="icon-filter"/>
                     </action>
-                    <action id="testcenter-search" name="Search" url="/taoProctoring/TestCenter/search" context="*"
+                    <action id="testcenter-search" name="Search" url="/taoProctoring/TestCenterManager/search" context="*"
                             group="search" binding="launchFinder">
                         <icon id="icon-find"/>
                     </action>
-                    <action id="testcenter-class-new" name="New class" url="/taoProctoring/TestCenter/addSubClass"
+                    <action id="testcenter-class-new" name="New class" url="/taoProctoring/TestCenterManager/addSubClass"
                             context="resource" group="tree" binding="subClass">
                         <icon id="icon-folder-open"/>
                     </action>
-                    <action id="testcenter-delete" name="Delete" url="/taoProctoring/TestCenter/delete"
+                    <action id="testcenter-delete" name="Delete" url="/taoProctoring/TestCenterManager/delete"
                             context="resource" group="tree" binding="removeNode">
                         <icon id="icon-bin"/>
                     </action>
-                    <action id="testcenter-move" name="Move" url="/taoProctoring/TestCenter/moveInstance"
+                    <action id="testcenter-move" name="Move" url="/taoProctoring/TestCenterManager/moveInstance"
                             context="instance" group="none" binding="moveNode">
                         <icon id="icon-move-item"/>
                     </action>
@@ -53,11 +53,11 @@
                             group="tree">
                         <icon id="icon-export"/>
                     </action>
-                    <action id="testcenter-duplicate" name="Duplicate" url="/taoProctoring/TestCenter/cloneInstance"
+                    <action id="testcenter-duplicate" name="Duplicate" url="/taoProctoring/TestCenterManager/cloneInstance"
                             context="instance" group="tree" binding="duplicateNode">
                         <icon id="icon-copy"/>
                     </action>
-                    <action id="testcenter-new" name="New Test Center" url="/taoProctoring/TestCenter/addInstance"
+                    <action id="testcenter-new" name="New Test Center" url="/taoProctoring/TestCenterManager/addInstance"
                             context="resource" group="tree" binding="instanciate">
                         <icon id="icon-users"/>
                     </action>
